3177 client onboarding analytics
