--- Modules/ClientAnalytics/Contracts/ClientAnalyticsServicesContract.php
@@ -1,7 +1,0 @@
-<?php
-namespace Modules\ClientAnalytics\Contracts;
-
-interface ClientAnalyticsServicesContract
-{
-
-}
