--- Modules/ClientAnalytics/Database/Seeders/ClientAnalyticsDatabaseSeeder.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace Modules\ClientAnalytics\Database\Seeders;
-
-use Illuminate\Database\Seeder;
-use Illuminate\Database\Eloquent\Model;
-
-class ClientAnalyticsDatabaseSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        Model::unguard();
-
-        // $this->call("OthersTableSeeder");
-    }
-}
