--- Modules/ClientAnalytics/Resources/views/layouts/master.blade.php
@@ -1,8 +1,0 @@
-@extends('layouts.app')
-@section('js_scripts')
-<script src="{{ mix('/js/clientanalytics.js') }}"></script>
-@endsection
-
-@section('css_scripts')
-<link href="{{ mix('/css/clientanalytics.css') }}" rel="stylesheet">
-@endsection
