--- Modules/ClientAnalytics/Routes/api.php
@@ -1,18 +1,0 @@
-<?php
-
-use Illuminate\Http\Request;
-
-/*
-|--------------------------------------------------------------------------
-| API Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register API routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| is assigned the "api" middleware group. Enjoy building your API!
-|
-*/
-
-Route::middleware('auth:api')->get('/clientanalytics', function (Request $request) {
-    return $request->user();
-});
