--- Modules/ClientAnalytics/Routes/web.php
@@ -1,17 +1,0 @@
-<?php
-
-/*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
-|
-*/
-
-Route::prefix('clientanalytics')->group(function() {
-    Route::get('/', 'ClientAnalyticsController@index')->name('ClientAnalytics.index');
-    
-});
