--- Modules/ClientAnalytics/composer.json
@@ -1,23 +1,0 @@
-{
-    "name": "nwidart/clientanalytics",
-    "description": "",
-    "authors": [
-        {
-            "name": "Nicolas Widart",
-            "email": "n.widart@gmail.com"
-        }
-    ],
-    "extra": {
-        "laravel": {
-            "providers": [],
-            "aliases": {
-
-            }
-        }
-    },
-    "autoload": {
-        "psr-4": {
-            "Modules\\ClientAnalytics\\": ""
-        }
-    }
-}
