--- Modules/ClientAnalytics/module.json
@@ -1,13 +1,0 @@
-{
-    "name": "ClientAnalytics",
-    "alias": "clientanalytics",
-    "description": "",
-    "keywords": [],
-    "order": 0,
-    "providers": [
-        "Modules\\ClientAnalytics\\Providers\\ClientAnalyticsServiceProvider"
-    ],
-    "aliases": {},
-    "files": [],
-    "requires": []
-}
