+++ Modules/Report/Routes/web.php
@@ -19,13 +19,14 @@ Route::prefix('report')->group(function () {
     Route::get('/delete/{id}', 'ReportController@delete')->name('report.delete');
     Route::post('/update/{id}', 'ReportController@update')->name('report.update');
     Route::get('/get-fte-report/{user}', 'UserReportController@getFteData')->name('reports.fte.get-report-data');
+    Route::get('/', 'ClientRevenueReportController@index')->name('reports.ClientOnboarding.index');
 
     Route::prefix('finance')->group(function () {
         Route::get('dashboard', 'FinanceReportController@dashboard')->name('reports.finance.dashboard');
         Route::get('get-report-data', 'FinanceReportController@getReportData')->name('reports.finance.get-report-data');
 
         Route::prefix('profit-and-loss')->group(function () {
-            Route::get('/', 'ProfitAndLossReportController@index')->name('reports.finance.profit-and-loss.index');
+            Route::get('/', 'ProfiClientOnboardingAnalyticstAndLossReportController@index')->name('reports.finance.profit-and-loss.index');
             Route::get('/detailed', 'ProfitAndLossReportController@detailed')->name('reports.finance.profit-and-loss.detailed');
             Route::get('profit-and-loss-report-export', 'ProfitAndLossReportController@profitAndLossReportExport')->name('reports.finance.profit-and-loss.report.export');
         });
